refactor(structure): rename store to stores
